add track_event
