--- lib/ahoy/intercom/client.rb
@@ -1,13 +1,0 @@
-require 'intercom'
-
-module Ahoy
-  module Intercom
-    module Client
-
-      def intercom
-        @intercom ||= ::Intercom::Client.new(token: ENV['INTERCOM_ACCESS_TOKEN'])
-      end
-      
-    end
-  end
-end
