+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 require "bundler/setup"
 require "ahoy/intercom"
 require "ahoy_matey"
+require "intercom"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -13,11 +14,11 @@ RSpec.configure do |config|
     c.syntax = :expect
   end
 
-  config.include Ahoy::Intercom::Client
-  config.include Ahoy::Intercom::Tracker
-
   config.before(:suite) do
-    class Ahoy::Store < Ahoy::Stores::BaseStore
+
+    Time.zone = 'UTC'
+
+    class Ahoy::Store < Ahoy::Intercom::Store
     end
   end
 end
